test(arel): fold predications-range.test.ts into predications.test.ts
